chore: consider docs-changes to be patch releases
Since our docstrings and our README.md files ship with the code,
we should publish a new version when these docs are updated.

This change will at least satisfy those requirements, and also possible
ship some new patch version for other docs that aren't visible in
the published artifact, but that sounds okay to me.

You can see on the commit-analyzer readme that the default `preset`
is `angular`[^1] (so no changes there, just making this explicit), and
explicit rules are additve to the base presets[^2].

[^1]: https://github.com/semantic-release/commit-analyzer#configuration
[^2]: https://github.com/semantic-release/commit-analyzer#releaserules