chore: remove explicit peer dependencies
We currently list monocle-ts and newtype-ts as explicit dependencies,
since these two packages are peer dependencies of io-ts-types[^1]
and npm 6 doesn't automatically install peer dependencies (like npm
7 does).

This commit removes the explicit peer dependencies, because when npm 7+
is used the behavior will be the same. Additionally, in npm 6 these
to packages won't be installed but they aren't needed, so there will
be no breaking change.

[^1]: https://github.com/gcanti/io-ts-types/blob/master/package.json#L40-L41