fix: check for compiler errors before emitting an openapi spec
Queries the typescript compiler for any errors before proceeding with
openapi generation. This catches issues such as `node_modules` not being
present, as it'll result in errors like "cannot resolve `io-ts`". It
also should preemptively catch other problems that we haven't
anticipated yet. There is a noticeable performance hit from this, but I
feel it is worth the trade-off for correctness.