chore: add GitHub Pages
To demo the site locally, run:

```
npm ci
npm --prefix packages/website start
```

This is a _heavy_ WIP but I'd also like to get this into a
continously-deploying state quickly, to prevent it from rotting too
long internally.

I plan to contribute regularly to these docs, so regardless of the
starting state we should have a good trajectory.