ci: use consistent concurrency limits
For release workflows:

- ensure there is only one workflow running concurrently
- do not skip any workflows

This is because releases act on shared resources (the deployed
environment, or git tags), so this concurrency limit acts as a mutex.

For CI (pre-merge) workflows:

- ensure only one workflow runs per branch
- cancel in-progress workflows

This is because we only care about the results of the latest workflows
to run, which act as required status checks. The impact is faster
reported status-check results, because of less queueing.

Ticket: 0