chore: update io-ts-types to 0.5.17
This version has a breaking change where the `nonEmptyArray` now returns
an actual `NonEmptyArray` type. This broke some tests but was easily
rectified. Aside from tests, the only type used from `io-ts-types` is
the `Json` one, so this shouldn't break consumers of this library.

This package is also used in `superagent-wrapper` and
`openapi-generator`, where it also shouldn't cause
backwards-compatibility issues.