feat: make `optionalized` strip explicit `undefined` values
`optionalized` determines which properties to make optional by checking
if the codec associated with the property accepts `undefined`. As a
result, it is impossible to use it to define types that have optional
properties whose type isn't a union containing `undefined`. When
`exactOptionalPropertyTypes` is `false` this isn't a big deal, because
all optional properties are implicitly treated this way. However, it
would be nice to be able to take advantage of the cleaner types enabled
by `exactOptionalPropertyTypes`.

This change offers a potential solution by changing the behavior of
`optionalized` such that it always deletes any properties that are
explicitly set to `undefined`. This is a breaking change, but seems
unlikely to cause issues in practice. If the previous behavior is
needed, it is always possible to use
`t.intersection([t.type(...),t.partial(...)])`.

This change also solves an inconsistency in `io-ts-http` exposed when
`exactOptionalPropertyTypes` is enabled. The `httpRequest` function is
implemented with `flatten`, which currently passes all keys to
`optionalized`. At the same time, `GenericHttpRequest` is defined using
`Record` types whose codomains do not include `undefined`. With
`exactOptionalPropertyTypes` enabled, this results in an error if, for
example, an optional query parameter is defined on an `httpRequest`
because of how `optionalized` forces all optional properties to be able
to be explicitly `undefined`. Rather than adding `t.undefined` to all
the records, this change presents an alternative that keeps the
`GenericHttpRequest` type as simple as possible.

BREAKING CHANGE: make `optionalized` delete any explicitly undefined
properties when encoding and decoding