test: decouple test case from typescript version
The externalModuleApiSpec tests are currently using the installed
TypeScript version as the api spec version. This means when we bump the
TypeScript version, we break tests.

This commit specifies a literal version number instead of pulling the
version number from TypeScript, so tests don't break when we update the
TypeScript version.