build: update patch-package patch file
The motivation behind this commit is to reduce the amount of
noise currently printed by `npm install` in the root of this
repository. Installing npm dependencies currently prints the following:

```
$ npm ci

> @api-ts/base@0.1.0 postinstall
> patch-package

patch-package 8.0.0
Applying patches...
@semantic-release/npm@10.0.5 ✔

Warning: patch-package detected a patch file version mismatch

  Don't worry! This is probably fine. The patch was still applied
  successfully. Here's the deets:

  Patch file created for

    @semantic-release/npm@10.0.5

  applied to

    @semantic-release/npm@12.0.1

  At path

    node_modules/@semantic-release/npm

  This warning is just to give you a heads-up. There is a small chance of
  breakage even though the patch was applied successfully. Make sure the package
  still behaves like you expect (you wrote tests, right?) and then run

    patch-package @semantic-release/npm

  to update the version in the patch file name and make this warning go away.
```

It looks like we created a patch for version
@semantic-release/npm@10.0.5, and have since updated (probably through
Dependabot) to @semantic-release/npm@12.0.1.

The underlying dependency doesn't seem to have changed in a way that
invalidates our patch, so I followed patch-package's instruction
and ran:

```sh
node_modules/.bin/patch-package @semantic-release/npm
```

This commit adds the generated file to version control. It also removes
the patch configured to run on @semantic-release/npm@10.0.5, since that
version is no longer installed in this repository (and removing this file
is necessary to silence the patch-package warning).

Update: and then I moved the comment over... you know, I could have just
renamed the file!